requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.30,<1.0
google-api-python-client>=2.170,<3.0
google-auth-oauthlib>=1.2,<2.0
